refactor: extract default monitors and simplify map creation in config
